LINA PUPPI- TINDER/puppi-tinder-django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
beautifulsoup4==4.12.2
requests==2.31.0